refactor(signup): created signupForm component + separated login css from signup css
