+++ src/components/SignupForm/SignupForm.tsx
@@ -1,0 +1,141 @@
+"use client";
+
+import Link from "next/link";
+
+import {ChangeEvent, FormEvent, useState} from "react";
+
+import {signupSchema} from "@/lib/schemas";
+import {Button, FormInput, PasswordInput} from "@/components/ui";
+
+import styles from "./SignupForm.module.css";
+
+const SignupForm = () => {
+
+    const [formData, setFormData] = useState({
+        email: "",
+        username: "",
+        firstName: "",
+        lastName: "",
+        password: "",
+        confirmPassword: ""
+    });
+
+    const [fieldErrors, setFieldErrors] = useState<Map<string, string>>(new Map());
+
+    const handleFormChange = (e: ChangeEvent<HTMLInputElement>) => {
+        setFormData({
+            ...formData,
+            [e.target.id]: e.target.value
+        });
+    };
+
+    const handleSubmit = (e: FormEvent) => {
+        e.preventDefault();
+
+        if (validateForm()) {
+
+        }
+    };
+
+    const validateForm = () => {
+        const newErrors: Map<string, string> = new Map<string, string>();
+
+        if (formData.password !== formData.confirmPassword) {
+            newErrors.set("confirmPassword", "Les mots de passe ne correspondent pas");
+        }
+
+        const validation = signupSchema.safeParse(formData);
+
+        if (!validation.success) {
+            validation.error.issues.forEach(issue => {
+                newErrors.set(issue.path[0] as string, issue.message);
+            });
+        }
+
+        setFieldErrors(newErrors);
+
+        return newErrors.size === 0;
+    };
+
+    return (
+        <div className={styles.formContainer}>
+            <div className={styles.formWrapper}>
+                <form onSubmit={handleSubmit} className={styles.form}>
+                    <FormInput
+                        label="Adresse e-mail"
+                        id="email"
+                        type="email"
+                        value={formData.email}
+                        onChange={handleFormChange}
+                        error={fieldErrors.get("email")}
+                        required
+                    />
+
+                    <FormInput
+                        label="Nom d'utilisateur"
+                        id="username"
+                        type="text"
+                        value={formData.username}
+                        onChange={handleFormChange}
+                        error={fieldErrors.get("username")}
+                        required
+                    />
+
+                    <div className="w-full flex justify-between gap-4">
+                        <FormInput
+                            label="Nom"
+                            id="lastName"
+                            type="text"
+                            value={formData.lastName}
+                            onChange={handleFormChange}
+                            error={fieldErrors.get("lastName")}
+                            required
+                        />
+
+                        <FormInput
+                            label="Prénom"
+                            id="firstName"
+                            type="text"
+                            value={formData.firstName}
+                            onChange={handleFormChange}
+                            error={fieldErrors.get("firstName")}
+                            required
+                        />
+                    </div>
+
+                    <PasswordInput
+                        label="Mot de passe"
+                        id="password"
+                        value={formData.password}
+                        onChange={handleFormChange}
+                        error={fieldErrors.get("password")}
+                        required
+                    />
+
+                    <PasswordInput
+                        label="Confirmer le mot de passe"
+                        id="confirmPassword"
+                        value={formData.confirmPassword}
+                        onChange={handleFormChange}
+                        error={fieldErrors.get("confirmPassword")}
+                        required
+                    />
+
+                    <Button type="submit" variant="primary" className={styles.signupButton}>
+                        S&apos;inscrire
+                    </Button>
+
+                    <div className={styles.loginLinkContainer}>
+                        <span className={styles.loginText}>Déjà un compte ? </span>
+                        <Link href="/login" className={styles.loginLink}>
+                            Se connecter
+                        </Link>
+                    </div>
+                </form>
+            </div>
+        </div>
+    );
+
+};
+
+export default SignupForm;
