+++ src/components/index.ts
@@ -1,5 +1,6 @@
 import CampaignCard from "@/components/CampaignCard/CampaignCard";
 import Pagination from "@/components/Pagination/Pagination";
 import SearchBar from "@/components/SearchBar/SearchBar";
+import SignupForm from "@/components/SignupForm/SignupForm";
 
-export {CampaignCard, Pagination, SearchBar};
+export {CampaignCard, Pagination, SearchBar, SignupForm};
